Add splitView layout screen to playground app
